scripts/requirements.txt
# Minimal requirements for dataset collection scripts
# Pin exact versions for reproducibility. Update cautiously.
datasets==2.20.0
pyyaml==6.0.2
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
tqdm==4.66.5
langdetect==1.0.9
datasketch==1.5.9
beautifulsoup4==4.12.3
requests==2.31.0
